Change Volk library to VMA (VulkanMemoryAllocator)
Performance may degrade slightly due to no longer using dynamically loaded function pointers during runtime which may incur a slight approx. up to 12% overhead /CPU/latency increase due to additional indirection/overhead/traversal of dlls/abstraction needed: https://gpuopen.com/learn/reducing-vulkan-api-call-overhead/

Remove VK_USE_PLATFORM_WIN32_KHR macro as by default in the main vulkan.h header it includes windows.h which seems to be a very slow header to parse and as we include only the mostly minimal header required/needed/necessitated anyway the whole/entire windows.h header is not needed and increased compilation time with little to no tangible benefits

+++ src/renderer2.hpp
@@ -36,7 +36,7 @@ struct renderer2// : Queues
   static constexpr void  setupRenderDraw() __attribute__( ( cold ) );
   [[gnu::pure, gnu::const, gnu::regcall, clang::preserve_all, gnu::hot]] void            drawFrame(const VkInit&, const SwapChain&, std::initializer_list<VkCommandBuffer>) const;
 
-  static constinit inline uint32_t               currentFrame;
+  static constinit inline uint8_t               currentFrame;
 
   constexpr static void memcpy2( __int256 *, __int256 const *, size_t ) __attribute__( ( __aligned__( 32 ), hot, flatten, preserve_all ) );
